### 2023-10-01 #5907 @editorialbot
[PRE REVIEW]: Stokesian Dynamics in Python
**Submitting author:** @Pecnut (<a href="http://orcid.org/0000-0003-1700-2873">Adam K. Townsend</a>)
**Repository:** https://github.com/Pecnut/stokesian-dynamics
**Branch with paper.md** (empty if default branch):
**Version:** 2.1.0
