### 2023-10-18 #5907 @philipcardiff
Hi @rajeshrinet, would you be interested in reviewing this submission? It seems similar to your [PyStokes package](https://github.com/rajeshrinet/pystokes).
