### 2023-10-19 #5907 @szwiep
@philipcardiff, thanks for the consideration. I’m comfortable reviewing the functionality and packaging of the project, but don’t think I have the appropriate amount of domain knowledge. Perhaps my co-author @smchartrand might be a better fit as a reviewer, if he has the time?
