### 2023-11-02 #6011 @editorialbot
[REVIEW]: Stokesian Dynamics in Python
**Submitting author:** @Pecnut (<a href="http://orcid.org/0000-0003-1700-2873">Adam K. Townsend</a>)
**Repository:** https://github.com/Pecnut/stokesian-dynamics
**Branch with paper.md** (empty if default branch):
**Version:** v3.0.2
