### 2023-11-04 #6011 @rajeshrinet
@philipcardiff, I have had my first pass on the software and the paper, as well as started my review checklist. I have added my first set of comments below. Once I hear back from the authors, I can proceed with my checklists.

## The paper
The paper is well written. It is thin on references to papers on Stokesian hydrodynamics. In addition, as per the review checklist, the `State of the field` section remains largely incomplete. There are a couple of libraries that exist to simulate Stokesian hydrodynamics of colloidal particles. These include:

* Hydrolib: https://doi.org/10.1016/0010-4655(95)00029-F
* libstokes: https://github.com/kichiki/libstokes
* Stokesdt: https://github.com/xing-liu/stokesdt
* PyStokes: https://github.com/rajeshrinet/pystokes [I must mention that I am involved in developing PyStokes.]

It is a requirement for the review that `authors describe how this software compares to other commonly-used packages.` I recommend suggesting in the paper the existence of related libraries through a section called `Related software.`

## The code

I could run files as suggested in the project README.md and generate outputs as indicated.

Here is a suggestion for presenting these results, which authors might consider. It would be helpful to have a comparison of the results of the simulation with standard results in the literature as a table displayed in the repo or through a Jupyter notebook (see also related comment below on using Jupyter notebooks).

For example, the authors could compare simulation results from their library along the lines of Table1 in this paper: https://doi.org/10.1017/S0022112088002411
I do understand that this is already covered in `example-c.py.` However, it is helpful to show the results directly on the GitHub page for potential users of the software package.

### Installation
There is no installation instruction provided for the library. I understand that the package files can be run directly into the correct folder. But it would be helpful to add in the readme:

`pip install -r requirements.txt`

Adding some information about installation in the README file is suggested.

### Testing
It needs to be clarified how to run tests to check if the library has been correctly installed (or downloaded). Indeed, a `test` folder exists. But no details on how to run tests are given in the README.

### Jupyter notebook examples

The authors can consider this suggestion as optional. It is helpful to have some example files in the format of jupyter notebooks (see https://jupyter.org/).

In particular, one can explain a bit of maths, run the code, and compare results to existing literature - all in the same window! Adding such notebooks will increase the reach of the software and make it more user-friendly.
