+++ joss/paper.bib
@@ -82,3 +82,121 @@
   abstract = {In many areas of suspension mechanics, such as filled polymer fluids or household products such as toothpaste, the suspending fluid itself is inherently non-Newtonian and may exhibit viscoelastic properties. In this paper, we extend the Stokesian Dynamics formalism to incorporate a simple model of viscoelasticity by using small spheres as `beads' in a bead\textendash spring dumbbell (such as is found in the derivation of Oldroyd and FENE constitutive models for dilute polymer solutions). Various different spring laws are then tested in both small-amplitude and large-amplitude oscillatory shear, and their rheological behaviour is compared to continuum constitutive models.},
   keywords = {Dumbbells,FENE force laws,LAOS,Oscillatory rheometry,SAOS,Simulations,Stokesian Dynamics},
 }
+
+@article{Townsend:2023,
+  title = {Generating, from scratch, the near-field asymptotic forms of scalar resistance functions for two unequal rigid spheres in low Reynolds number flow},
+  author = {Townsend, A.~K.},
+  year = {2023},
+  month = dec,
+  journal = {Physics of Fluids},
+  volume = {35},
+  number = {12},
+  pages = {127126},
+  abstract = "{The motion of rigid spherical particles suspended in a low Reynolds number fluid can be related to the forces, torques, and stresslets acting upon them by 22 scalar resistance functions, commonly notated     X  11 A ,    X  12 A ,    Y  11 A, etc. Near-field asymptotic forms of these resistance functions were derived in Jeffrey and Onishi [“Calculation of the resistance and mobility functions for two unequal rigid spheres in low-Reynolds-number flow,” J. Fluid Mech. 139, 261–290 (1984)] and Jeffrey [“The calculation of the low Reynolds number resistance functions for two unequal spheres,” Phys. Fluids A 4, 16–29 (1992)]; these forms are now used in several numerical methods for suspension mechanics. However, the first of these important papers contains a number of small errors, which make it difficult for the reader to correctly evaluate the functions for parameters not explicitly tabulated. This short article comprehensively corrects these errors and adds formulas that were originally omitted from both papers, so that the reader can verify and implement the equations independently. The corrected expressions, rationalized and using contemporary nondimensionalization, are shown to match mid-field values of these scalars, which are calculated through an alternative method. A Python script to generate and evaluate these functions is provided.}",
+  issn = {1070-6631},
+  doi = {10.1063/5.0175697},
+  url = {https://doi.org/10.1063/5.0175697},
+  eprint = {https://pubs.aip.org/aip/pof/article-pdf/doi/10.1063/5.0175697/18277079/127126\_1\_5.0175697.pdf},
+}
+
+@book{Ladyzhenskaya:1964,
+  title = {The Mathematical Theory of Viscous Incompressible Flow},
+  author = {Ladyzhenskaya, O. A.},
+  year = {1964},
+  month = jan,
+  edition = {revised English},
+  publisher = {{Gordon and Breach}},
+  langid = {english},
+}
+
+@article{Durlofsky:1987,
+  title = {Dynamic Simulation of Hydrodynamically Interacting Particles},
+  author = {Durlofsky, L. and Brady, J. F. and Bossis, G.},
+  year = {1987},
+  journal = {Journal of Fluid Mechanics},
+  volume = {180},
[…]
