+++ joss/paper.md
@@ -27,6 +27,7 @@ In the most basic case, Stokes’ law states that a single sphere of radius $a$,
 Stokesian Dynamics, at its heart, is an extension of this linear relationship between the force acting on a particle and the velocity at which it travels. As a method, it is adaptable and continues to be used in the field, providing some interesting insight into the behaviour of particle suspensions. Validations with experiments have shown it to provide results within acceptable error.
 
 The Stokesian Dynamics recipe can be summarised as follows:
+
 * Compute long-range interactions between particles. This is done by using a truncated expansion of the boundary integral equation for unbounded Stokes flow (@Ladyzhenskaya:1964, @Durlofsky:1987). For periodic domains, this is done by using the Ewald summation method (@Ewald:1921, @Brady:1988).
 * Compute short-range lubrication between particles. This is done by interpolating pre-computed results on a pairwise basis (@Jeffrey:1984, @Jeffrey:1992, @Kim:2005, @Wilson:2013, @Townsend:2023).
 * Combining the two.
