### 2024-01-12 #6011 @Pecnut
Hello everyone,

Thank you @rajeshrinet for your comments. I hope I have addressed them in the latest commit. In particular:

* The state of the field / related software is much richer for your recommendations. I am particularly impressed by your library! I hope you agree with my summaries of the other available software.
* The basics of Stokesian Dynamics is better cited now. I suspect that people won't come to this code base if they don't already know what Stokesian Dynamics is, but I think this will definitely help.
* Installation information is now explicit.
* Testing information is also now explicit.
* Examples comparing against the literature are now much more explicitly signposted in the README, including images which show the output of the example scripts. These images include the data from figures in the literature we are comparing against.
* I also think these graphs are better demonstrations of literature validation than tables of values, given that the data isn't explicitly given in the source papers, only the graphs.
* I agree that Jupyter Notebooks are nice, but given that this software doesn't behave as a drop-in Python library, I don't think that it would work here (I did try!). For example, there are a number of examples where input files have to be edited, and this isn't the natural function of a Jupyter Notebook. Instead, I have tried to improve the README so that the examples/lit validation examples are better signposted and easier to follow.

I enjoyed making these improvements and I hope you agree they are indeed improvements! Thank you @rajeshrinet.

@szwiep I look forward to hearing your thoughts.
