### 2024-02-05 #6011 @szwiep
Hi @peacnut. First off, this is really exciting software, even more so because you've decided to make it open source. I've listed my comments below. Many are simply suggestions and I'm happy to discuss any of them further with you.

### Errors

- The `resource` package is a Linux specific package. When running the first sample run on a Windows 10 system, I get the error listed below. I was able to resolve the error by commenting out lines 21 and 738 in `run_simulation.py`
```
H:\stokesian-dynamics (master -> origin)
(.venv) λ python run_simulation.py 2 1 0.5 1
Traceback (most recent call last):
  File "H:\stokesian-dynamics\run_simulation.py", line 21, in <module>
	import resource
ModuleNotFoundError: No module named 'resource'
```
- Possible bad string formulation on line 960 in `run_simulation.py`. strftime('%A %-d %B %Y, %H:%M') should be strftime('%A **%d** %B %Y, %H:%M'). Throws an error on Windows 10 systems but not Linux.
- For any of the predefined set ups that included more than one particle, particles were plotted outside the view of the Figure. Occurs on both Windows 10 and Linux. Is this expected behaviour? Here are some examples:
![c1](https://github.com/openjournals/joss-reviews/assets/57023024/d13f0823-4b82-476f-8226-1317de5c0020)
![c2](https://github.com/openjournals/joss-reviews/assets/57023024/7b57997b-bed4-48a0-b562-f82995977ffd)

- "**Note**: as received, the software is only able to perform simulations with particles of size ratio 1:1, 1:10 and 1:100. To increase this, see section 10." Should reference section 11, not 10.
- I receive a [Failed to send email] output at the end of each of my runs even though I have send_email = False.

### Environment

- Requiring the exact Python version 3.9.13 is quite restrictive. It's usually safe to say your software is compatible within the minor version e.g 3.9. If you'd like to be a bit more robust, you can set up [GitHub Actions](https://docs.github.com/en/actions/automating-builds-and-tests/building-and-testing-python) to automatically run your tests across multiple Python versions every time you make a PR.
- I'd suggest adding package versions (e.g "numpy == 1.26.0") to your requirements.txt. This helps ensure users will be using the same dependency versions as you even as the development of those packages progress with time.

### Code Organization

- I think re-organizing the structure of your project could make navigating the code/documentation more intuitive. I tend to loosely follow the [Hitchhiker's Guide](https://docs.python-guide.org/writing/structure/) when structuring projects. You'll see this structure reflected in many (most?) repos, so following the core ideas of it can help users through familiarity.
	- For example, one possible structure could look like:
```
stokesian-dynamics
│   README.md
│   requirements.txt
│   ...
└───stokesian-dynamics
│   │   run_simulation.py
│   │   inputs.py
│   └───functions
│   |   │   email.py
│   |   │   generate_Minfinity.py
│   |   │   ...
|   └───convert
│   |   │   convert_resistance_scalars_to_d_form.py
│   |   │   ...
|   └───setups
│   |   │   positions_setups.py
│   |   │   input_setups.py
|   └───plotting
│       │   plot_particle_positions.py
│       │   plot_particle_positions_video.py
│       │   ...
└───docs
|    │  non_README_docs_here.txt
|    |  ...
└───tests
|   │   ...
...
```
### Documentation

- I'm not sure if I've missed it somewhere, but there should ideally be stand-alone documentation of all the parameters that the model accepts. The document should provide descriptions of the parameters, what their defaults are, and acceptable ranges of inputs.
- It would be helpful to have some images/screenshots of the plots so that users know what to expect from your plotting feature.

### Packaging

- Have you considered packaging stokesian-dynamics and hosting it on PyPI? I think your software could be a great candidate for this by setting up a [console script entry point](https://setuptools.pypa.io/en/latest/userguide/entry_point.html)!
