### 2024-02-07 cf5181d
Error fixes from JOSS review + GIFs for examples
