### 2024-02-07 b21ab4a
Fix width of images in README
