### 2024-02-07 028a5d8
Package versions added to requirements
