+++ functions/timestepping.py
@@ -7,15 +7,15 @@ import numba
 import time
 import datetime
 from numpy import linalg
-from functions_generate_grand_resistance_matrix import (
+from functions.generate_grand_resistance_matrix import (
     generate_grand_resistance_matrix, generate_grand_resistance_matrix_periodic)
-from functions_shared import (posdata_data, format_elapsed_time, throw_error,
+from functions.shared import (posdata_data, format_elapsed_time, throw_error,
                               shear_basis_vectors)
-from functions_simulation_tools import (
+from functions.simulation_tools import (
     construct_force_vector_from_fts, deconstruct_velocity_vector_for_fts,
     fts_to_fte_matrix, fte_to_ufte_matrix, ufte_to_ufteu_matrix,
     fts_to_duf_matrix)
-from input_setups import input_ftsuoe
+from setups.inputs import input_ftsuoe
 
 
 def euler_timestep(x, u, timestep):
