### 2024-02-09 9840aa5
Structural changes better in line with Hitchhiker's Guide
