+++ stokesian_dynamics/setups/functions_positions.py
@@ -1,0 +1,413 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+# Adam Townsend, adam@adamtownsend.com, 08/02/2024
+
+"""Useful distribution functions for setting up initial positions of particles.
+"""
+
+import numpy as np
+from os import sys
+
+
+def random_point_in_box(random_box_bottom_left, random_box_top_right):
+    """Return random coordinate inside a box."""
+    dimensions = (np.array(random_box_top_right)
+                  - np.array(random_box_bottom_left))
+    return random_box_bottom_left + np.random.rand(3)*dimensions
+
+
+def overlaps(pos1, pos2, radius1, radius2):
+    """Do objects 1 and 2 overlap each other?"""
+    return np.linalg.norm(pos1-pos2) < (radius1+radius2)
+
+
+def overlaps_periodic(pos1, pos2, radius1, radius2, dxy):
+    """Do objects 1 and 2 overlap each other, when separated by dxy?"""
+    return np.linalg.norm(pos1-pos2-dxy) < (radius1+radius2)
+
+
+def randomise_spheres(random_box_bottom_left, random_box_top_right,
+                      random_sphere_sizes, current_sphere_sizes,
+                      current_sphere_positions):
+    """Place spheres randomly inside a box, avoiding currently
+    placed spheres.
+    """
+    sphere_positions_out = current_sphere_positions
+    num_current_spheres = current_sphere_sizes.shape[0]
+    all_sphere_sizes = np.concatenate([current_sphere_sizes,
+                                       random_sphere_sizes])
+    print("Randomly distributing " + str(len(random_sphere_sizes))
+          + " spheres (using a naive method in Python)... ", end=" ")
+    for i in range(len(random_sphere_sizes)):
+        while 1 == 1:
+            proposed_sphere_position = random_point_in_box(
+                random_box_bottom_left, random_box_top_right)
+            too_close = 0
+            for j in range(num_current_spheres+i):
+                if overlaps(proposed_sphere_position, sphere_positions_out[j],
+                            random_sphere_sizes[i], all_sphere_sizes[j]):
+                    too_close = too_close + 1
+            if too_close == 0:
+                symbol = '|'
+                if (i+1) % 10 == 0:
+                    symbol = 'X'
+                if (i+1) % 100 == 0:
+                    symbol = 'C'
+                sys.stdout.write(symbol)
+                sys.stdout.flush()
+                break
+        sphere_positions_out = np.append(sphere_positions_out,
+                                         [proposed_sphere_position], axis=0)
+    print(" succeeded.")
+
+    random_sphere_positions = sphere_positions_out[
+        current_sphere_positions.shape[0]:all_sphere_sizes.shape[0]]
+    distance_matrix = np.linalg.norm(
+        random_sphere_positions-random_sphere_positions[:, None], axis=2)
+    min_element_distance = np.min(distance_matrix[np.nonzero(distance_matrix)])
+    two_closest_elements = np.where(distance_matrix == min_element_distance)
+    print("Min added sphere s': " + str(min_element_distance/random_sphere_sizes[0]))
+    print("Closest two spheres: " + str(two_closest_elements[0]))
+
+    box_dimensions = abs(np.asarray(random_box_top_right)
+                         - np.asarray(random_box_bottom_left))
+
+    if box_dimensions[1] == 0:  # 2D
+        box_volume = box_dimensions[0]*box_dimensions[2]
+        sphere_volumes = np.pi*np.dot(random_sphere_sizes, random_sphere_sizes)
+    else:  # 3D
[…]
