+++ inputs.py
@@ -1,262 +1,0 @@
-#!/usr/bin/env python
-# -*- coding: utf-8 -*-
-# Adam Townsend, adam@adamtownsend.com, 07/06/2017
-
-import sys  # Overrides from command line at the bottom of this script
-import glob
-import numpy as np
-from numba import config
-from setups.positions import pos_setup
-from functions.shared import add_sphere_rotations_to_positions, throw_error
-
-
-# --------------------
-# PHYSICAL INPUTS
-setup_number = 1
-input_number = 1
-fully_2d_problem = False
-
-# Timestep (default = 0.1)
-timestep = 0.04
-
-# Number of frames to run simulation for
-num_frames = 100
-
-# How to solve the equation U=MF. {'fts','fte','ufte','ufteu','duf'}
-input_form = 'fte'
-extract_force_on_wall_due_to_dumbbells = False
-
-# --------------------------------------------------|
-# OVERRIDES FROM COMMAND LINE
-args = sys.argv[1:]
-number_of_args = len(args)
-running_on_legion = 0
-
-try:
-    if number_of_args >= 6:
-        extract_force_on_wall_due_to_dumbbells = (
-            args[5] in ["True", "true", "t", "T", "1"])
-    if number_of_args >= 5:
-        input_form = args[4]
-    if number_of_args >= 4:
-        num_frames = int(args[3])
-    if number_of_args >= 3:
-        timestep = float(args[2])
-    if number_of_args >= 2:
-        input_number = int(args[1])
-    if number_of_args >= 1:
-        setup_number = int(args[0])
-except ValueError:
-    throw_error("Invalid input. Command-line input should come in the "
-                + "order: setup number, input number, timestep, no. frames, "
-                + "input form, extract dumbbell force Boolean.")
-# --------------------------------------------------|
-
-
-posdata, setup_description = pos_setup(setup_number)
-
-# Checkpointing ---------|
-
-# Checkpoint override. If a file exists with the same input number, position
-# number, number of frames and timestep, AND it ends in _TEMP, try to
-# continue it.
-
-# Initialise variables
-checkpoint_filename = ''
-checkpoint_start_from_frame = 0
-
-# NOTE: comment this section to disable Checkpointing
-# Otherwise read in list of temporary files
-checkpoint_start_from_frame = 0
-for filename in sorted(glob.glob("output/*_TEMP.npz"), reverse=True):
-    checkpoint_setup_number = int(filename.split("-")[1][1:])
-    checkpoint_input_number = int(filename.split("-")[2][1:])
-    checkpoint_num_frames = int(filename.split("-")[3][:-2])
-    if filename.split("-")[4][-1] == "e":
-        raw_timestep = filename.split("-")[4][1:] + "-" + filename.split("-")[5]
-    else:
-        raw_timestep = filename.split("-")[4][1:]
-    checkpoint_timestep = float(raw_timestep.replace("p", "."))
-    if (checkpoint_setup_number == setup_number
-        and checkpoint_input_number == input_number
-        and checkpoint_num_frames == num_frames
-            and checkpoint_timestep == timestep):
-        ignored_posdata, setup_description = pos_setup(checkpoint_setup_number)
-        try:
-            with np.load(filename) as data1:
-                positions_centres = data1['centres']
-                positions_deltax = data1['deltax']
-            num_particles = positions_centres.shape[1]
-            num_dumbbells = positions_deltax.shape[1]
-            num_spheres = num_particles - num_dumbbells
-            sphere_positions = positions_centres[-1, 0:num_spheres, :]
[…]
