+++ stokesian_dynamics/README.md
@@ -1,0 +1,5 @@
+# Stokesian Dynamics in Python
+
+This folder contains the main files for the Stokesian Dynamics software.
+
+Please read the [main repository README](../README.md) in the parent directory for all information and guidance for how to use the software.
