### 2024-02-09 52aa688
Bump scipy from 1.9.1 to 1.11.1
Bumps [scipy](https://github.com/scipy/scipy) from 1.9.1 to 1.11.1.
- [Release notes](https://github.com/scipy/scipy/releases)
- [Commits](https://github.com/scipy/scipy/compare/v1.9.1...v1.11.1)

---
updated-dependencies:
- dependency-name: scipy
  dependency-type: direct:production
...

Signed-off-by: dependabot[bot] <support@github.com>
