+++ requirements.txt
@@ -2,4 +2,4 @@ matplotlib==3.5.2
 numba==0.55.1
 numpy==1.21.5
 pytest==7.1.2
-scipy==1.9.1
+scipy==1.11.1
