+++ requirements.txt
@@ -1,5 +1,5 @@
 matplotlib==3.5.2
 numba==0.55.1
-numpy==1.21.5
+numpy==1.22.0
 pytest==7.1.2
 scipy==1.9.1
