+++ stokesian_dynamics/setups/tests/inputs.py
@@ -2,7 +2,7 @@
 # -*- coding: utf-8 -*-
 # Adam Townsend, adam@adamtownsend.com, 07/02/2024
 
-"""Numbered list of all input forces/velocities used in the tests in the 
+"""Numbered list of all input forces/velocities used in the tests in the
 /test/ folder."""
 
 import numpy as np
