+++ setup.cfg
@@ -2,4 +2,4 @@
 max-line-length = 120
 min-version = 3.0
 # TODO: remove these ignores
-ignore = E231, E303, W503, E226, W504, E741, E722, E305, F401
+ignore = E231, E303, W503, E226, W504, E741, E722, E305, F401, E402
