### 2024-02-09 #6011 @Pecnut
Hi @szwiep! Thank you for your thorough and extremely useful comments. I have enjoyed thinking about them and acting upon them.

**Errors**:

* All errors re Windows should now be fixed: sorry, I didn't have a Windows machine to test it on. Let me know if the latest version doesn't run properly.
* Re overly large graphics: The size of the viewing box for the graphical output has been set to a better default. It is also now made clear in the README how to change this to something appropriate for each test case.
* Wrong section number now fixed.
* Email warning now fixed as well.

**Environment**:

* I agree re Python 3.9.13: I was just trying to say that I had tested it on this version. I've changed the text to say "3.9" and GitHub Actions is now set up to test it on version 3.8+.
* I have added package versions to requirements.txt but currently I have added them with `>=` rather than `==`. Do you think `==` is better? On the one hand, I see that `>=` sets me up to fail if something changes in the packages over time. But doesn't `==` require the user to potentially download an old version of a package? (Also I noticed that the version of Numba that I wanted to ask for isn't actually available any more – if availability of specific packages changes then this also seems like a problem).

**Code organisation**:

I have really enjoyed thinking about this – thank you for your suggestion and for linking the Hitchhiker's Guide. I have made pretty substantial changes to the structure that I really think improve the project. Let me know what you think.

In particular, because the code isn't set up as a package, there are a couple of awkward compromises that I've made:

* The location of the main README is, I think, still good, but it's a bit less obvious which directory should be your working directory to run the codes now. I have tried to be as clear as possible.
* I've had to use `sys.path.append( )` in a couple of places to import files correctly (tests and plotting). I think this is the neatest solution.

**Documentation**:

* I have created standalone documentation for all of the possible settings for the simulation software. This is in `docs/sd-settings.md`.
* I have added more images and GIFs in the README to show what to expect when you see a simulation running.

**Packaging**

Yes I have thought about this! Thank you for suggesting PyPI and console script entry points. Right now I think packaging this up will take a bit more time than I have available, particularly because it will require more thought about structure (and more learning/education about packages in general on my end). I would like to take a look at this but more as a medium-term project (so not right now).

Thank you again @szwiep for your time and your comments! I think the software is much better for it.
