### 2024-02-10 10b1795
New example of mixed forces/velocities
