### 2024-02-17 #6011 @szwiep
Hi @pecnut, thank you for thoughtfully considering my comments. Your updates look great!

## New Comments
- Looks `psutil` is missing from your requirements.txt. On both Linux and Windows systems, I run into `ModuleNotFoundError: No module named 'psutil'` when running the tool in a fresh environment.

## Follow-up Comments
- Plotting, email, and resource package errors are all resolved in Windows/Linux.
- The restructuring of your project looks great. I think your compromises between script and package structuring are well executed, and this sets you up well for packaging your tool in the future.
- I think your addition of GIFs to the README is an especially helpful tool for users!

> Do you think `==` is better? On the one hand, I see that `>=` sets me up to fail if something changes in the packages over time. But doesn't `==` require the user to potentially download an old version of a package?
- This is a good question and I suspect the answer is a bit subjective. Generally, downloading a package that is a bit older than the most recent stable release is fine and gives you the peace of mind that all dependencies work as expected. But it is also good to allow users the flexibility to use newer/different versions of packages. A compromise could be [`~=`](https://stackoverflow.com/questions/39590187/in-requirements-txt-what-does-tilde-equals-mean), which allows new package versions but only as long as they are backward-compatible with your listed version. Either way, as long as the version that you developed the tool with is listed, then choosing `==`, `>=`, or `~=`, shouldn't be of much consequence. Users can always backtrack to those versions if something goes awry!

@philipcardiff once the `ModuleNotFoundError` error is resolved, my review is complete.
