+++ requirements.txt
@@ -1,5 +1,6 @@
-matplotlib>=3.5.2
-numba>=0.55.1
-numpy>=1.21.5
-pytest>=7.1.2
-scipy>=1.9.1
+matplotlib~=3.5.2
+numba~=0.55.1
+numpy~=1.21.5
+psutil~=5.9.0
+pytest~=7.1.2
+scipy~=1.9.1
