### 2024-02-19 f3c909a
Revert "Requirements to ~="
This reverts commit 6c73575696256b5073bd7e507ac34671b61d0d8c.
