+++ requirements.txt
@@ -1,5 +1,6 @@
-matplotlib>=3.5.2
-numba>=0.55.1
-numpy>=1.21.5
-pytest>=7.1.2
-scipy>=1.9.1
+matplotlib~=3.8.0
+numba~=0.59.0
+numpy~=1.26.3
+psutil~=5.9.0
+pytest~=7.4.0
+scipy~=1.11.4
