### 2024-02-19 #6011 @Pecnut
Thank you @szwiep! I am much happier with the structure of the project now and I think the improved README will definitely help.

* I have added `psutil` to `requirements.txt` so `ModuleNotFoundError` should now be fixed (sorry, an oversight).
* I have gone for the `~=` approach you suggest, but also increased the package versions in line with the system I am testing the project on locally (since one of the versions I was using was no longer available).

Hopefully that should all be sorted, then!

@philipcardiff Over to you, I think!
