### 2024-02-19 d88280e
README tweaks: remove footnotes, add psutil
