### 2024-02-20 #6011 @philipcardiff
I proposed minor edits to the paper at https://github.com/Pecnut/stokesian-dynamics/pull/10
