+++ docs/joss/paper.bib
@@ -3,7 +3,7 @@
     title = {The mechanics of suspensions},
     author = {Townsend, A.~K.},
     year = {2017},
-    school = {University College London}    
+    school = {University College London}
 }
 
 @article{Brady:1988,
@@ -184,6 +184,7 @@
   publisher = {{Dover Publications}},
   address = {{Mineola, NY, USA}},
   isbn = {0-486-44219-5 978-0-486-44219-8},
+  doi = {10.1016/C2013-0-04644-0},
   langid = {english}
 }
 
