### 2024-02-20 e768e82
Bibliography additional information
