+++ docs/joss/paper.bib
@@ -1,7 +1,7 @@
 @phdthesis{Townsend:2017,
   	url = {https://discovery.ucl.ac.uk/id/eprint/1559911/},
     title = {The mechanics of suspensions},
-    author = {Townsend, A.~K.},
+    author = {Townsend, A. K.},
     year = {2017},
     school = {University College London}
 }
@@ -9,7 +9,7 @@
 @article{Brady:1988,
   url = {https://www.annualreviews.org/doi/abs/10.1146/annurev.fl.20.010188.000551},
   title = {Stokesian {{Dynamics}}},
-  author = {Brady, J.~F. and Bossis, G.},
+  author = {Brady, J. F. and Bossis, G.},
   year = {1988},
   journal = {Annual Review of Fluid Mechanics},
   volume = {20},
@@ -22,7 +22,7 @@
 
 @article{Townsend:2018b,
   title = {Anomalous Effect of Turning off Long-Range Mobility Interactions in {{Stokesian}} Dynamics},
-  author = {Townsend, A.~K. and Wilson, H.~J.},
+  author = {Townsend, A. K. and Wilson, H. J.},
   year = {2018},
   month = jul,
   journal = {Physics of Fluids},
@@ -37,7 +37,7 @@
 @article{Townsend:2017c,
   title = {Frictional Shear Thickening in Suspensions: {{The}} Effect of Rigid Asperities},
   shorttitle = {Frictional Shear Thickening in Suspensions},
-  author = {Townsend, A.~K. and Wilson, H.~J.},
+  author = {Townsend, A. K. and Wilson, H. J.},
   year = {2017},
   month = nov,
   journal = {Physics of Fluids},
@@ -52,7 +52,7 @@
 
 @article{Townsend:2017b,
   title = {Simulations of a Heavy Ball Falling through a Sheared Suspension},
-  author = {Townsend, A.~K. and Wilson, H.~J.},
+  author = {Townsend, A. K. and Wilson, H. J.},
   year = {2017},
   month = dec,
   journal = {Journal of Engineering Mathematics},
@@ -70,7 +70,7 @@
 
 @article{Townsend:2018a,
   title = {Small- and Large-Amplitude Oscillatory Rheometry with Bead\textendash Spring Dumbbells in {{Stokesian Dynamics}} to Mimic Viscoelasticity},
-  author = {Townsend, A.~K. and Wilson, H.~J.},
+  author = {Townsend, A. K. and Wilson, H. J.},
   year = {2018},
   month = nov,
   journal = {Journal of Non-Newtonian Fluid Mechanics},
@@ -84,8 +84,8 @@
 }
 
 @article{Townsend:2023,
-  title = {Generating, from scratch, the near-field asymptotic forms of scalar resistance functions for two unequal rigid spheres in low Reynolds number flow},
-  author = {Townsend, A.~K.},
+  title = {Generating, from scratch, the near-field asymptotic forms of scalar resistance functions for two unequal rigid spheres in low {{Reynolds}} number flow},
+  author = {Townsend, A. K.},
   year = {2023},
   month = dec,
   journal = {Physics of Fluids},
@@ -104,9 +104,12 @@
   author = {Ladyzhenskaya, O. A.},
   year = {1964},
   month = jan,
+  series = {Mathematics and Its {{Applications}}},
   edition = {revised English},
+  volume = {2},
   publisher = {{Gordon and Breach}},
   langid = {english},
+  annotation = {Translated from the Russian by R. A. Silverman}
 }
 
 @article{Durlofsky:1987,
