### 2024-02-20 0fc1641
Bibliography edit
