+++ docs/joss/paper.bib
@@ -106,7 +106,7 @@
   month = jan,
   series = {Mathematics and Its {{Applications}}},
   edition = {revised English},
-  volume = {2},
+  number = {2},
   publisher = {{Gordon and Breach}},
   langid = {english},
   annotation = {Translated from the Russian by R. A. Silverman}
