### 2024-02-20 af1a2db
DOI and JOSS badges in README
