### 2024-02-20 b9e6aa3
Semantic change on 'while True'
