### 2024-02-20 e3f1018
Unused arguments in plot_all_spheres
