### 2024-02-20 8094b3e
Unused arguments in plot_all_dumbbells
