### 2024-02-20 03c746b
Unused arguments in plotting
