### 2024-02-20 d2b7804
Unused arguments in generate_output_FTSUOE
