### 2024-02-20 0963d1b
Make clearer the purpose of local flag
