### 2024-02-20 44c99e6
Removed unused arguments from plotting functions
