+++ stokesian_dynamics/functions/graphics.py
@@ -92,9 +92,8 @@ def between0and1(num):
         return num
 
 
-def plot_sphere(ax, frameno, viewbox_bottomleft_topright, position,
-                previous_position, trace_paths, radius, sphere_rotations,
-                dumbbell=0, sphere_colour='b'):
+def plot_sphere(ax, frameno, position, previous_position, trace_paths, radius,
+                sphere_rotations, dumbbell=0, sphere_colour='b'):
     n = normal(ax)
 
     if dumbbell == 0:
@@ -151,8 +150,7 @@ def plot_sphere(ax, frameno, viewbox_bottomleft_topright, position,
     return (p, line1, line2, linetrace)
 
 
-def plot_dumbbell(ax, viewbox_bottomleft_topright, position, trace_paths,
-                  radius, dx, rot_theta=0, rot_phi=0,
+def plot_dumbbell(ax, position, trace_paths, radius, dx, rot_theta=0, rot_phi=0,
                   dumbbell_colour=['b', 'b'], no_line=False):
     pos1 = position - 0.5*dx
     pos2 = position + 0.5*dx
@@ -185,9 +183,8 @@ def plot_dumbbell(ax, viewbox_bottomleft_topright, position, trace_paths,
     return (p1, p2) if no_line else (p1, p2, line)
 
 
-def plot_all_spheres(ax, frameno, viewbox_bottomleft_topright, posdata,
-                     previous_step_posdata, trace_paths, sphere_trace_lines,
-                     f_spheres):
+def plot_all_spheres(ax, frameno, posdata, previous_step_posdata, trace_paths,
+                     sphere_trace_lines, f_spheres):
     (sphere_sizes, sphere_positions, sphere_rotations, dumbbell_sizes,
      dumbbell_positions, dumbbell_deltax) = posdata
     (previous_sphere_sizes, previous_sphere_positions,
@@ -200,7 +197,7 @@ def plot_all_spheres(ax, frameno, viewbox_bottomleft_topright, posdata,
         C = between0and1(np.linalg.norm(np.array(f_spheres[i], float)))
         sphere_colour = [C, 0, 1-C]
         (p, l1, l2, ltrace) = plot_sphere(
-            ax, frameno, viewbox_bottomleft_topright, sphere_positions[i, :],
+            ax, frameno, sphere_positions[i, :],
             previous_sphere_positions[i, :], trace_paths, sphere_sizes[i],
             sphere_rotations[i], sphere_colour=sphere_colour)
         spheres.append(p)
@@ -210,10 +207,8 @@ def plot_all_spheres(ax, frameno, viewbox_bottomleft_topright, posdata,
     return spheres, sphere_lines, sphere_trace_lines
 
 
-def plot_all_dumbbells(ax, frameno, viewbox_bottomleft_topright, posdata,
-                       previous_step_posdata, trace_paths,
-                       dumbbell_trace_lines, f_dumbbells, deltaf_dumbbells,
-                       max_DFb_out=1, no_line=False):
+def plot_all_dumbbells(ax, posdata, trace_paths, dumbbell_trace_lines,
+                       f_dumbbells, deltaf_dumbbells, no_line=False):
     (sphere_sizes, sphere_positions, sphere_rotations, dumbbell_sizes,
      dumbbell_positions, dumbbell_deltax) = posdata
     dumbbell_lines = []
@@ -242,14 +237,14 @@ def plot_all_dumbbells(ax, frameno, viewbox_bottomleft_topright, posdata,
         dumbbell_colour = [dumbbell_colour1, dumbbell_colour2]
         if no_line:
             sphere1, sphere2 = plot_dumbbell(
[…]
