### 2024-02-20 ef007e9
Removed unused plotting arguments
