### 2024-02-20 0d9881f
Removal of unused variables
