### 2024-02-21 9969f39
Removed unused code
