### 2024-02-21 #6011 @philipcardiff
> ⚠️ [Error preparing paper acceptance](https://github.com/openjournals/joss-papers/actions/runs/7988562826?check_suite_focus=true). The generated XML metadata file is invalid.
>
> ```yaml
> ID ref-BradyU003A1988 already defined
> ```

Hi @Pecnut, It seems that you have two instances of `Brady:1988` in your bib file; can you check this? Thanks.
