+++ docs/joss/paper.bib
@@ -6,7 +6,7 @@
     school = {University College London}
 }
 
-@article{Brady:1988,
+@article{Brady:1988a,
   url = {https://www.annualreviews.org/doi/abs/10.1146/annurev.fl.20.010188.000551},
   title = {Stokesian {{Dynamics}}},
   author = {Brady, J. F. and Bossis, G.},
@@ -123,7 +123,7 @@
   langid = {english},
 }
 
-@article{Brady:1988,
+@article{Brady:1988b,
   title = {Dynamic Simulation of Hydrodynamically Interacting Suspensions},
   author = {Brady, John F. and Phillips, Ronald J. and Lester, Julia C. and Bossis, Georges},
   year = {1988},
