### 2024-02-21 a6bba4c
Bibliography formatting
