+++ docs/joss/paper.bib
@@ -125,7 +125,7 @@
 
 @article{Brady:1988b,
   title = {Dynamic Simulation of Hydrodynamically Interacting Suspensions},
-  author = {Brady, John F. and Phillips, Ronald J. and Lester, Julia C. and Bossis, Georges},
+  author = {Brady, J. F. and Phillips, R. J. and Lester, J. C. and Bossis, G.},
   year = {1988},
   month = oct,
   journal = {Journal of Fluid Mechanics},
@@ -182,7 +182,7 @@
 @book{Kim:2005,
   title = {Microhydrodynamics: Principles and Selected Applications},
   shorttitle = {Microhydrodynamics},
-  author = {Kim, Sangtae and Karrila, Seppo J.},
+  author = {Kim, S. and Karrila, S. J.},
   year = {2005},
   publisher = {{Dover Publications}},
   address = {{Mineola, NY, USA}},
@@ -193,7 +193,7 @@
 
 @article{Wilson:2013,
   title = {Stokes Flow Past Three Spheres},
-  author = {Wilson, Helen J.},
+  author = {Wilson, H. J.},
   year = {2013},
   month = jul,
   journal = {Journal of Computational Physics},
