### 2024-02-21 #6011 @kyleniemeyer
Thanks @philipcardiff. No new release is needed if only the paper changed, since JOSS archives the paper itself.
