### 2024-02-22 c35ef94
Bibliography style fixes
