+++ docs/joss/paper.bib
@@ -204,3 +204,54 @@
   urldate = {2014-05-27},
   langid = {english},
 }
+
+@misc{Swan:2013,
+  title = {Stokesian {{Dynamics}}, {{live}}},
+  author = {Swan, J. W.},
+  year = {2013},
+  urldate = {2024-02-22},
+  howpublished = {http://web.mit.edu/swangroup/sd-game.shtml}
+}
+
+@article{Hydrolib:1995,
+  title = {{{HYDROLIB}}: A Library for the Evaluation of Hydrodynamic Interactions in Colloidal Suspensions},
+  shorttitle = {{{HYDROLIB}}},
+  author = {Hinsen, K.},
+  year = {1995},
+  month = aug,
+  journal = {Computer Physics Communications},
+  volume = {88},
+  number = {2},
+  pages = {327--340},
+  issn = {0010-4655},
+  doi = {10.1016/0010-4655(95)00029-F}
+}
+
+@misc{libstokes:1999,
+  title = {libstokes},
+  author = {Ichiki, K.},
+  year = {1999},
+  urldate = {2024-02-22},
+  abstract = {library for Stokesian dynamics},
+  copyright = {GPL-2.0}
+}
+
+@misc{StokesDT:2015,
+  title = {{{StokesDT}}},
+  author = {Liu, X. and Chow, E.},
+  year = {2015}
+}
+
+@article{PyStokes:2020,
+  title = {{{PyStokes}}: Phoresis and {{Stokesian}} Hydrodynamics in {{Python}}},
+  shorttitle = {{{PyStokes}}},
+  author = {Singh, Rajesh and Adhikari, Ronojoy},
+  year = {2020},
+  month = jun,
+  journal = {Journal of Open Source Software},
+  volume = {5},
+  number = {50},
+  pages = {2318},
+  issn = {2475-9066},
+  doi = {10.21105/joss.02318}
+}
