### 2024-02-22 58f6725
Bibliography style fix
