+++ docs/joss/paper.bib
@@ -210,7 +210,7 @@
   author = {Swan, J. W.},
   year = {2013},
   urldate = {2024-02-22},
-  howpublished = {http://web.mit.edu/swangroup/sd-game.shtml}
+  howpublished = {\url{http://web.mit.edu/swangroup/sd-game.shtml}}
 }
 
 @article{Hydrolib:1995,
@@ -233,13 +233,17 @@
   year = {1999},
   urldate = {2024-02-22},
   abstract = {library for Stokesian dynamics},
-  copyright = {GPL-2.0}
+  copyright = {GPL-2.0},
+  journal = {GitHub repository},
+  howpublished = {\url{https://github.com/kichiki/libstokes}}
 }
 
 @misc{StokesDT:2015,
   title = {{{StokesDT}}},
   author = {Liu, X. and Chow, E.},
-  year = {2015}
+  year = {2015},
+  journal = {GitHub repository},
+  howpublished = {\url{https://github.com/xing-liu/stokesdt}}
 }
 
 @article{PyStokes:2020,
