+++ docs/joss/paper.bib
@@ -234,7 +234,7 @@
   urldate = {2024-02-22},
   abstract = {library for Stokesian dynamics},
   copyright = {GPL-2.0},
-  journal = {GitHub repository},
+  publisher = {GitHub},
   howpublished = {\url{https://github.com/kichiki/libstokes}}
 }
 
@@ -242,7 +242,7 @@
   title = {{{StokesDT}}},
   author = {Liu, X. and Chow, E.},
   year = {2015},
-  journal = {GitHub repository},
+  publisher = {GitHub},
   howpublished = {\url{https://github.com/xing-liu/stokesdt}}
 }
 
