### 2024-02-22 876fb49
Force lowercase on bib item
