+++ docs/joss/paper.bib
@@ -228,7 +228,7 @@
 }
 
 @misc{libstokes:1999,
-  title = {libstokes},
+  title = {{libstokes}},
   author = {Ichiki, K.},
   year = {1999},
   urldate = {2024-02-22},
