### 2024-02-22 #6011 @Pecnut
Hi @kyleniemeyer – I think I've addressed all these things now. Thank you for checking!

1. Yes, this should just be me. @philipcardiff was indeed a (very good) paper correction.
2. Fixed
3. Fixed
